Moved App.cities to App.data.cities
